fix: some mixins
