+++ src/main/java/net/ccbluex/liquidbounce/injection/mixins/minecraft/entity/MixinPlayerLikeEntity.java
@@ -1,0 +1,44 @@
+/*
+ * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
+ *
+ * Copyright (c) 2015 - 2025 CCBlueX
+ *
+ * LiquidBounce is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LiquidBounce is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.ccbluex.liquidbounce.injection.mixins.minecraft.entity;
+
+import net.ccbluex.liquidbounce.features.module.modules.movement.ModuleNoPose;
+import net.minecraft.entity.EntityDimensions;
+import net.minecraft.entity.EntityPose;
+import net.minecraft.entity.PlayerLikeEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(PlayerLikeEntity.class)
+public class MixinPlayerLikeEntity {
+    /**
+     * Sneak height fix
+     * @see net.ccbluex.liquidbounce.features.module.modules.movement.ModuleNoPose
+     */
+    @Inject(method = "getBaseDimensions", at = @At("RETURN"), cancellable = true)
+    private void hookGetBaseDimensions(EntityPose pose, CallbackInfoReturnable<EntityDimensions> cir) {
+        if (pose == EntityPose.CROUCHING) {
+            var dimensions = ModuleNoPose.INSTANCE.modifySneakHeight();  /* If module/setting is not enabled, modifySneakHeight() returns null */
+            if (dimensions != null) cir.setReturnValue(dimensions);
+        }
+    }
+}
