+++ src/main/resources/liquidbounce.mixins.json
@@ -59,7 +59,6 @@
     "minecraft.entity.MixinMobEntity",
     "minecraft.entity.MixinMobSpawnerBlockEntityRenderer",
     "minecraft.entity.MixinPlayerEntity",
-    "minecraft.entity.MixinPlayerEntityAccessor",
     "minecraft.entity.projectile.MixinFireworkRocketEntity",
     "minecraft.entity.projectile.MixinFishingBobberEntity",
     "minecraft.fluid.MixinFlowableFluid",
@@ -146,7 +145,8 @@
     "truffle.MixinHostClassLoader",
     "truffle.MixinHostContext",
     "truffle.MixinNpmCompatibleESModuleLoader",
-    "truffle.MixinTruffleLanguage"
+    "truffle.MixinTruffleLanguage",
+    "minecraft.entity.MixinPlayerLikeEntity"
   ],
   "server": [],
   "injectors": {
